requirements.txt
boto3==1.37.5
botocore==1.37.5
jmespath==1.0.1
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
s3transfer==0.11.3
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
